% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_halo_brackets}
\alias{read_halo_brackets}
\title{Read in professional Halo bracket data}
\source{
\url{https://liquipedia.net}
}
\usage{
read_halo_brackets()
}
\description{
Read in professional Halo bracket data
}
