% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_halo_gamertags}
\alias{read_halo_gamertags}
\title{Read in professional Halo team data}
\source{
\url{https://halodatahive.com}
}
\usage{
read_halo_gamertags()
}
\description{
Read in professional Halo team data
}
