% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_halo_players}
\alias{read_halo_players}
\title{Read in professional Halo player data}
\source{
\url{https://liquipedia.net}
}
\usage{
read_halo_players()
}
\description{
Read in professional Halo player data
}
