% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_halo_rosters}
\alias{read_halo_rosters}
\title{Read in professional Halo roster data}
\source{
\url{https://liquipedia.net}
}
\usage{
read_halo_rosters()
}
\description{
Read in professional Halo roster data
}
