% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_halo_tournaments}
\alias{read_halo_tournaments}
\title{Read in professional Halo tournament data}
\source{
\url{https://liquipedia.net}
}
\usage{
read_halo_tournaments()
}
\description{
Note that this CSV may not cover all the tournaments implied by the brackets
}
