% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{update_data}
\alias{update_data}
\title{Update all data}
\usage{
update_data(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite all existing data}

\item{scrape_time}{Time that data was last scraped}
}
\description{
Update all data
}
