% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{update_data}
\alias{update_data}
\title{Update all data}
\usage{
update_data(scrape_time = lubridate::now(), overwrite = FALSE)
}
\arguments{
\item{scrape_time}{Time that data was last scraped}

\item{overwrite}{Whether to overwrite all existing data}
}
\description{
Update all data
}
